+++ constants/colors.ts
@@ -1,0 +1,11 @@
+const colors = {
+  base: '#ffffff',
+  background: '#f5f5f5',
+  primary: '#f5f6fa',
+  accent: '#7983ff',
+  'main-text': '#2b2b2b',
+  'secondary-text': '#b1b0b4',
+  'accent-text': '#3d5afe',
+}
+
+export default colors
